requirements.txt
fastapi
uvicorn
python-multipart
pydantic
httpx
python-dotenv
celery
redis
google-generativeai